ai-services/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
cachetools